+++ lib/src/dart_code_generator/dart_code_generator.dart
@@ -191,6 +191,48 @@ class {{NAME}} {
   List<String> _generateMethodToJson(TypeDeclaration type) {
     const template = '''
   Map<String, dynamic> toJson() {
+{{STATEMENT}}
+  }''';
+
+    var result = template;
+    var properties = type.properties;
+    var names = properties.values.map((e) => e.name);
+    var statement = <String>[];
+    if (names.isEmpty) {
+      statement.add('    return {};');
+    } else {
+      statement.add('    return {');
+      for (var name in names) {
+        var property = properties[name];
+        var alias = property.alias;
+        if (alias != null) {
+          alias = alias.replaceAll('\$', '\\\$');
+          alias = alias.replaceAll('\'', '\\\'');
+        } else {
+          alias = name;
+        }
+
+        var writer = _getWriter(name, property.type, true);
+        var sb = StringBuffer();
+        sb.write('      ');
+        sb.write('\'');
+        sb.write(alias);
+        sb.write('\': ');
+        sb.write(writer);
+        sb.write(',');
+        statement.add(sb.toString());
+      }
+
+      statement.add('    };');
+    }
+
+    result = result.replaceAll('{{STATEMENT}}', statement.join('\n'));
+    return LineSplitter().convert(result).toList();
+  }
+
+  List<String> _generateMethodToJson_(TypeDeclaration type) {
+    const template = '''
+  Map<String, dynamic> toJson() {
     var result = <String, dynamic>{};
 {{STATEMENTS}}
     return result;